Made GVNS an own module, correction and extension wrt. initial solution/construction heuristics
